order

+++ insert_order_data.py
@@ -32,7 +32,7 @@ for mid in materials:
         'physical_stock': random.randint(100, 10000),
         'available_stock': random.randint(100, 10000),
         'allocated_stock': random.randint(100, 10000),
-        'pending_outbound': random.randint(100, 10000)
+        'pending_outbound':random.randint(100, 10000)
     })
 write_csv('instance\Material.csv', list(material_data[0].keys()), material_data)
 
@@ -122,7 +122,7 @@ for i, qid in enumerate(quotation_ids[:NUM_RECORDS]):
         'order_date': (today + timedelta(days=i+2)).strftime('%Y-%m-%d'),
         'required_delivery_date': (today + timedelta(days=i+10)).strftime('%Y-%m-%d'),
         'status': random.choice(['草稿','已创建','已审核','部分发货','已完成','已取消']),
-        'total_amount': quotation_data[i]['total_amount'],
+        'total_amount': float(quotation_data[i]['total_amount']),
         'credit_check_result': random.choice(['通过', '待审核']),
         'remarks': fake.sentence()
     })
@@ -132,8 +132,8 @@ write_csv('instance\SalesOrder.csv', list(sales_order_data[0].keys()), sales_ord
 order_item_data = []
 for i, oid in enumerate(order_ids):
     for item_no in range(1, 4):
-        qty = random.randint(10, 200)
-        price = round(random.uniform(30, 250), 2)
+        qty = int(random.randint(10, 200))
+        price = float(round(random.uniform(30, 250), 2))
         order_item_data.append({
             'sales_order_id': oid,
             'item_no': item_no,
